+++ src/main/web/scoreEntry/scoreEntryModule.js
@@ -61,14 +61,47 @@ const score_entry_module = {}
      * Shows the dialog named "upload-dialog" and closes it when done and loads select_team.jsp.
      */
     score_entry_module.uploadScore = function(storageData) {
-        //FIXME try creating dialog dynamically
-        alert("FIXME need to submit data here");
-        
-        
-            // FIXME: show wait dialog
-            // FIXME: send AJAX request
-            // FIXME: on response show message dialog and redirect to select_team.jsp
-            // FIXME: on success deleteStorageData
-    }
+        const dialog = document.createElement("div");
+        dialog.classList.add("fll-sw-ui-dialog");
+
+        const dialogContainer = document.createElement("div");
+        dialog.appendChild(dialogContainer);
+
+        const textElement = document.createElement("div");
+        dialogContainer.appendChild(textElement);
+        textElement.innerHTML = "Please wait while the score is uploaded...";
+
+        const closeButton = document.createElement("button");
+        closeButton.setAttribute("type", "button");
+        closeButton.innerText = "Close";
+        closeButton.classList.add("fll-sw-ui-inactive");
+        closeButton.addEventListener("click", () => {
+            window.location = "select_team.jsp";
+        });
+        dialogContainer.appendChild(closeButton);
+
+        document.body.appendChild(dialog);
+
+        // make the map look like an object so that it parses into a Java Map
+        const dataToUpload = JSON.stringify(Object.fromEntries(storageData.formValues));
+        fetch("SubmitScoreEntry", {
+            method: "POST",
+            headers: { 'Content-Type': 'application/json' },
+            body: dataToUpload
+        }).then(checkJsonResponse).then(function(result) {
+            if (result.success) {
+                textElement.innerHTML = result.message;
+
+                // clear the data as it was successfully submitted
+                score_entry_module.deleteStorageData(storageData);
+            } else {
+                textElement.innerHTML = result.message;
+            }
+        }).catch(function(result) {
+            textElement.innerHTML = "<div class='error'>Error talking to the server. The score data is stored locally. Error: " + result.message + "</div>";
+        }).then(function() {
+            closeButton.classList.remove("fll-sw-ui-inactive");
+        });
+    };
 
 }